refactor(visual-regions): split form candidates
Visual-region construction now has separate modules for form, caption, label, and ruled-table concerns. Moving form candidate clustering out of the central builder keeps form-region behavior easier to review while leaving the visual-region output unchanged.
